chore: add serialization for groth16
